requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.3
playwright==1.51.0
beautifulsoup4==4.13.3